requirements.txt
Django
djangorestframework
mongoengine
django-rest-framework-mongoengine
celery
redis
dopy==0.3.5
boto3
